requirements.txt
streamlit
python-dotenv
langchain>=0.3.0
langchain-openai
langchain-pinecone
langchain-community
langchain-text-splitters
pinecone-client
pypdf
openai>=1.0
